requirements.txt
fastapi
uvicorn
python-multipart
pandas
google-generativeai